Début du déplacement de la vignette de groupe actif vers le haut
Suite : la rendre horizontale; on harmonise ainsi mobile et bureau; on
permet sur mobile de voir le score de la conf en direct quand on fait
son test; on garde le menu simple, non mouvant;
